Re-design Names and IDs
At present we have both names and ids for resources and rules. Names are mainly for short descriptions and ids are for identifiers.
When creating a rule or resource, the user provides a name, and then rule-engine returns a server-side-generated id.

If a user run `emqx_ctl rules create <name> <config>` multiple times using the same name, multiple rules (with different ids) would be created by the rule-engine. This may confuse users, as they would expect that names can be used as identifers.

The problem is, `name` is not actually needed, it is only a word describing what the rule is for. We have a `descr` field for that. Users should create a rule by `emqx_ctl rules create <config>` and get an id, that's it.

The changes in this PR:

- Remove the field `name` from resources and rules.
  As explained above, we only need IDs.
- Change format of action-name from `<resource_type>:<name>` to `<name>`.
  Don't add a prefix on action names, use exactly the one defined by resource providers.
  We already have a `type` field for actions.
- Change format of rule-id from `<rule-name>:<timestamp>` to `rule:<short-uuid>`.
  e.g. rule:3bc859b2